requirements.txt
langchain
transformers
torch
pillow
langchain-ollama
streamlit
